+++ Source/Csla.Shared/Rules/BusinessRuleManager.cs
@@ -8,7 +8,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace Csla.Rules
 {
@@ -17,7 +16,6 @@ namespace Csla.Rules
   /// </summary>
   public class BusinessRuleManager
   {
-#if !(ANDROID || IOS) && !NETFX_CORE
     private static Lazy<System.Collections.Concurrent.ConcurrentDictionary<RuleSetKey, BusinessRuleManager>> _perTypeRules =
       new Lazy<System.Collections.Concurrent.ConcurrentDictionary<RuleSetKey, BusinessRuleManager>>();
 
@@ -28,38 +26,6 @@ namespace Csla.Rules
       var key = new RuleSetKey { Type = type, RuleSet = ruleSet };
       return _perTypeRules.Value.GetOrAdd(key, (t) => { return new BusinessRuleManager(); });
     }
-#else
-    private static Dictionary<RuleSetKey, BusinessRuleManager> _perTypeRules = new Dictionary<RuleSetKey, BusinessRuleManager>();
-
-    internal static BusinessRuleManager GetRulesForType(Type type, string ruleSet)
-    {
-      if (ruleSet == ApplicationContext.DefaultRuleSet) ruleSet = null;
-
-      BusinessRuleManager result = null;
-      var key = new RuleSetKey { Type = type, RuleSet = ruleSet };
-      var found = false;
-      try
-      {
-        found = _perTypeRules.TryGetValue(key, out result);
-      }
-      catch
-      { /* failure will drop into !found block */ }
-      if (!found)
-      {
-        lock (_perTypeRules)
-        {
-          if (!_perTypeRules.TryGetValue(key, out result))
-          {
-            result = new BusinessRuleManager();
-            _perTypeRules.Add(key, result);
-          }
-        }
-      }
-      return result;
-    }
-
-#endif
-
 
     /// <summary>
     /// Remove/delete all the rules for the given type.
@@ -69,29 +35,16 @@ namespace Csla.Rules
     {
       lock (_perTypeRules)
       {
-
         // the first RuleSet is already added to list when this check is executed so so if count > 1 then we have already initialized type rules.
-#if !(ANDROID || IOS) && !NETFX_CORE
         var typeRules = _perTypeRules.Value.Where(value => value.Key.Type == type);
         foreach (var key in typeRules)
-        {
-          BusinessRuleManager manager;
-          _perTypeRules.Value.TryRemove(key.Key, out manager);
-        }
-#else
-        var typeRules = _perTypeRules.Where(value => value.Key.Type == type).ToArray();
-        foreach (var key in typeRules)
-        {
-          _perTypeRules.Remove(key.Key);
-        }
-#endif
+          _perTypeRules.Value.TryRemove(key.Key, out BusinessRuleManager manager);
       }
     }
 
     internal static BusinessRuleManager GetRulesForType(Type type)
     {
       return GetRulesForType(type, null);
-
     }
 
     private class RuleSetKey
@@ -101,8 +54,7 @@ namespace Csla.Rules
 
       public override bool Equals(object obj)
       {
-        var other = obj as RuleSetKey;
-        if (other == null)
+        if (!(obj is RuleSetKey other))
           return false;
         else
           return this.Type.Equals(other.Type) && RuleSet == other.RuleSet;
@@ -117,7 +69,7 @@ namespace Csla.Rules
     /// <summary>
     /// Gets the list of rule objects for the business type.
     /// </summary>
-    public List<IBusinessRule> Rules { get; private set; }
+    public List<IBusinessRuleBase> Rules { get; private set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether the rules have been
@@ -127,7 +79,7 @@ namespace Csla.Rules
 
     private BusinessRuleManager()
     {
-      Rules = new List<IBusinessRule>();
+      Rules = new List<IBusinessRuleBase>();
     }
   }
 }